requirements.txt
python-telegram-bot==22.1
dotenv==0.9.9
huggingface==0.0.1
huggingface-hub==0.32.2
newspaper4k==0.9.3.1
joblib==1.5.1
lxml_html_clean==0.4.2
numpy==1.26.4
openai==1.82.1
orjson==3.10.18
pandas==2.1.4
pinecone[asyncio]==7.0.2
nltk==3.9.1
scikit-learn==1.5.1
scipy==1.15.3
spacy==3.8.7
SQLAlchemy==2.0.41
toolz==1.0.0
torch==2.7.0
transformers==4.52.3
typing-extensions
uk_core_news_lg @ https://github.com/explosion/spacy-models/releases/download/uk_core_news_lg-3.8.0/uk_core_news_lg-3.8.0-py3-none-any.whl#sha256=60b56f8fa5525925cf8e6266a80e907431c3d475f874c75c68b62f3e12620b0b
